+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Adding `SparseSet`.
 * Adding iterator methods to `BitSet`.
+* `BitSet` now relies on `Uint32Array` rather than `Uint8Array`.
 
 ## 0.13.0